+++ src/main/java/pl/pjatk/learnease/entity/learn/Subject.java
@@ -1,7 +1,10 @@
 package pl.pjatk.learnease.entity.learn;
 
 import jakarta.persistence.*;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Entity
 @Table(name = "subjects")
@@ -13,7 +16,7 @@ public class Subject {
     @Id
     @Column(name = "subject_id", unique = true)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer subjectId;
+    private Long subjectId;
 
     @Column(name = "name", nullable = false)
     private String name;
