+++ src/main/java/pl/pjatk/learnease/entity/quiz/QuizParticipant.java
@@ -19,7 +19,7 @@ public class QuizParticipant {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "participant_id", nullable = false)
-    private Integer participantId;
+    private Long participantId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "quiz_room_id", nullable = false)
