+++ src/main/java/pl/pjatk/learnease/entity/ranking/Point.java
@@ -1,7 +1,10 @@
 package pl.pjatk.learnease.entity.ranking;
 
 import jakarta.persistence.*;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import pl.pjatk.learnease.entity.learn.Material;
 import pl.pjatk.learnease.entity.test.CheckingTest;
 import pl.pjatk.learnease.entity.user.User;
@@ -16,7 +19,7 @@ public class Point {
     @Id
     @Column(name = "point_id", unique = true)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer pointId;
+    private Long pointId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "checking_test_id")
