+++ src/main/java/pl/pjatk/learnease/entity/ranking/Ranking.java
@@ -1,7 +1,10 @@
 package pl.pjatk.learnease.entity.ranking;
 
 import jakarta.persistence.*;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import pl.pjatk.learnease.entity.user.User;
 
 import java.time.LocalDateTime;
@@ -16,13 +19,13 @@ public class Ranking {
     @Id
     @Column(name = "ranking_id", unique = true)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer rankingId;
+    private Long rankingId;
 
     @Column(name = "rank", nullable = false)
-    private Integer rank;
+    private Long rank;
 
     @Column(name = "previous_rank")
-    private Integer previousRank;
+    private Long previousRank;
 
     @Column(name = "points", nullable = false)
     private Float points;
