+++ src/main/java/pl/pjatk/learnease/entity/user/Level.java
@@ -1,7 +1,10 @@
 package pl.pjatk.learnease.entity.user;
 
 import jakarta.persistence.*;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Entity
 @Table(name = "levels")
@@ -13,7 +16,7 @@ public class Level {
     @Id
     @Column(name = "level_id", unique = true)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer levelId;
+    private Long levelId;
 
     @Column(name = "name", nullable = false)
     private String name;
